+++ client/app/scripts/controllers/products/products-controller.js
@@ -8,18 +8,11 @@
  * Controller of the gemStoreApp
  */
 angular.module('gemStoreApp')
-	.controller('ProductsCtrl', ['$scope', '$state', 'productsService', function ($scope, $state, productsService) {
+	.controller('ProductsCtrl', ['$scope', '$state', '$http', 'productsService', function ($scope, $state, $http, productsService) {
 
 		$scope.products = {};
 		console.log('Query the database for products!');
 		$scope.products = productsService.query();
 
-		//$scope.orderProp = "name"
-	}])
-
-	.controller('ProductDetailCtrl', ['$scope', '$state', '$stateParams', 'productsService', function ($scope, $state, $stateParams, productsService) {
-
-		$scope.product = [];
-		console.log('Query the database for product with the id of ' + $stateParams.id + '!');
-		$scope.product = productsService.get({id: $stateParams.id});
+		//$scope.orderProp = "name"		
 	}]);