+++ client/app/scripts/services/products/products-service.js
@@ -1,7 +1,7 @@
 angular.module('gemStoreApp.productService',['ngResource'])
 
     .factory('productsService', function($resource) {
-        return $resource('/products/:id', {},{
+        return $resource('/products/:id', {id:'@id'},{
         	'update': { method: 'PUT'}
         });
     });